Add `LoadPizzak` service to load pizzas for selection
This update includes a new `loadPizzak` service which generates a list of pizzas and allows fetching this list. Main class is also updated to include getter for this new service. Additionally, the way pizzas are added in the `rendelesPage` is now changed to use a dropdown select of available pizzas, replacing the old individual input fields for pizza details. It also adds a new `@types/node` dev dependency.